requirements.txt
semantic-kernel>=0.4.0
python-dotenv>=1.0.0 